% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM.csv}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance}
  \item{DAMAGE}{Damage}
  ...
}
}
\source{
<FIREDAM.csv>
}
\usage{
fire
}
\description{
Damages caused by fires I think
}
\keyword{datasets}
