% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{p}{}
}
\value{
barplot
}
\description{
Binomial simulation
}
\examples{
mybin(iter=1000,n=10,p=0.5)
}
