% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{numeric vector}

\item{x}{numeric vector}

\item{fun}{string}

\item{alpha}{numeric vector}

\item{cx}{numeric vector}

\item{...}{}
}
\value{
invisible list
}
\description{
myboot2
}
\examples{
myboot2(iter=10000, x, fun="var")
}
