% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{mycltu}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{numeric vector}

\item{iter}{numeric vector}

\item{a}{numeric vector}

\item{b}{numeric vector}
}
\value{
curves
}
\description{
mycltu
}
\examples{
mycltu(n=20,iter=100000)

}
