% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{numeric vector}

\item{mu}{numeric vector}

\item{sig}{numeric vector}

\item{...}{other}
}
\value{
list
}
\description{
mymlnorm
}
\examples{
mymlnorm(x=c(10, 12, 14, 13, 12, 11, 13, 14), mu = 1:20, sig= 1:10)
}
