% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{numeric vector}

\item{sigma}{numeric vector}

\item{a}{numeric vector}
}
\value{
curve and list
}
\description{
myncurve
}
\examples{
myncurve(mu=10,sigma=2, a=6)
}
