% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{numeric vector}

\item{gamma}{numeric vector}

\item{p}{numeric vector}
}
\value{
list
}
\description{
ntickets
}
\examples{
ntickets(N=400, gamma=0.02, p=0.95)
}
